hello-agent-assignment/packaged-agent/pyproject.toml
[project]
name = "packaged-agent"
version = "0.1.0"
description = "A packaged agent example using OpenAI API directly."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.40.0",
    "python-dotenv>=1.0.1",
]

[project.scripts]
packaged-agent = "packaged_agent.main:run"

[build-system]
requires = ["uv_build>=0.8.11,<0.9.0"]
build-backend = "uv_build"

[tool.uv.workspace]
members = [
    "packaged-agent",
]